+++ indico/web/client/js/react/components/StickyWithScrollBack.jsx
@@ -5,16 +5,32 @@
 // modify it under the terms of the MIT License; see the
 // LICENSE file for more details.
 
-import React, {useState} from 'react';
+import React, {useState, useEffect} from 'react';
 import PropTypes from 'prop-types';
 import {Sticky} from 'semantic-ui-react';
 
 import {ScrollButton} from 'indico/react/components';
+import {useResponsive} from 'indico/react/util';
 
 import './StickyWithScrollBack.module.scss';
 
-export default function StickyWithScrollBack({children, context}) {
+const MIN_RESPONSIVE_OFFSET = 200;
+
+export default function StickyWithScrollBack({children, context, responsive}) {
   const [scrollButtonVisible, setScrollButtonVisible] = useState(false);
+  const [responsiveScrollOffset, setResponsiveScrollOffset] = useState(window.pageYOffset);
+
+  const {isPhone, isTablet, isLandscape} = useResponsive();
+  const isResponsiveDevice = responsive && (isPhone || isTablet) && isLandscape;
+
+  useEffect(() => {
+    window.addEventListener('scroll', onScroll, false);
+    return () => window.removeEventListener('scroll', onScroll, false);
+  });
+
+  function onScroll() {
+    setResponsiveScrollOffset(window.pageYOffset);
+  }
 
   return (
     <Sticky
@@ -22,9 +38,15 @@ export default function StickyWithScrollBack({children, context}) {
       styleName="sticky-content"
       onStick={() => setScrollButtonVisible(true)}
       onUnstick={() => setScrollButtonVisible(false)}
+      active={!isResponsiveDevice}
     >
       {children}
-      <ScrollButton visible={scrollButtonVisible} />
+      <ScrollButton
+        visible={
+          scrollButtonVisible ||
+          (isResponsiveDevice && responsiveScrollOffset > MIN_RESPONSIVE_OFFSET)
+        }
+      />
     </Sticky>
   );
 }
@@ -32,9 +54,11 @@ export default function StickyWithScrollBack({children, context}) {
 StickyWithScrollBack.propTypes = {
   children: PropTypes.node,
   context: PropTypes.object,
+  responsive: PropTypes.bool,
 };
 
 StickyWithScrollBack.defaultProps = {
   children: null,
   context: null,
+  responsive: false,
 };