+++ indico/web/client/js/react/util/Responsive.jsx
@@ -7,7 +7,7 @@
 
 import PropTypes from 'prop-types';
 import React from 'react';
-import Responsive from 'react-responsive';
+import Responsive, {useMediaQuery} from 'react-responsive';
 
 const DIMENSIONS = {
   tablet: 768,
@@ -15,10 +15,15 @@ const DIMENSIONS = {
   wideScreen: 1200,
 };
 
-const factory = (minDimension, maxDimension) => {
+const ORIENTATIONS = {
+  landscape: 'landscape',
+  portrait: 'portrait',
+};
+
+const dimensionFactory = (minDimension, maxDimension) => {
   /**
    * This component extends `Responsive` from `react-responsive`, adding some
-   * useful configuration options.
+   * useful dimension configuration options.
    */
   function _SizeSpec({andLarger, andSmaller, orElse, children, onlyIf, ...restProps}) {
     return (
@@ -54,9 +59,58 @@ const factory = (minDimension, maxDimension) => {
   return _SizeSpec;
 };
 
+const orientationFactory = orientation => {
+  /**
+   * This component extends `Responsive` from `react-responsive`, adding some
+   * useful orientation configuration options.
+   */
+  function _OrientationSpec({orElse, children, onlyIf, ...restProps}) {
+    return (
+      <Responsive orientation={orientation} {...restProps}>
+        {matches => (matches && onlyIf ? children : orElse)}
+      </Responsive>
+    );
+  }
+
+  _OrientationSpec.propTypes = {
+    /** allows for negative cases to be specified easily */
+    orElse: PropTypes.node,
+    /** allows adding an extra condition to render the matching content */
+    onlyIf: PropTypes.bool,
+    children: PropTypes.node.isRequired,
+  };
+
+  _OrientationSpec.defaultProps = {
+    onlyIf: true,
+    orElse: null,
+  };
+
+  return _OrientationSpec;
+};
+
 export default Object.assign(Responsive, {
-  WideScreen: factory(DIMENSIONS.wideScreen, null),
-  Desktop: factory(DIMENSIONS.computer, DIMENSIONS.wideScreen),
-  Tablet: factory(DIMENSIONS.tablet, DIMENSIONS.computer),
-  Phone: factory(null, DIMENSIONS.tablet),
+  WideScreen: dimensionFactory(DIMENSIONS.wideScreen, null),
+  Desktop: dimensionFactory(DIMENSIONS.computer, DIMENSIONS.wideScreen),
+  Tablet: dimensionFactory(DIMENSIONS.tablet, DIMENSIONS.computer),
+  Phone: dimensionFactory(null, DIMENSIONS.tablet),
+  Portrait: orientationFactory(ORIENTATIONS.portrait),
+  Landscape: orientationFactory(ORIENTATIONS.landscape),
 });
+
+export function useResponsive() {
+  const belowWideScreen = DIMENSIONS.wideScreen - 1;
+  const belowComputingScreen = DIMENSIONS.computer - 1;
+  const belowTabletScreen = DIMENSIONS.tablet - 1;
+  return {
+    isWideScreen: useMediaQuery({query: `(min-width: ${DIMENSIONS.computer}px)`}),
+    isDesktop: useMediaQuery({
+      query: `(min-width: ${DIMENSIONS.computer}px) and (max-width: ${belowWideScreen}px)`,
+    }),
+    isTablet: useMediaQuery({
+      query: `(min-width: ${DIMENSIONS.tablet}px) and (max-width: ${belowComputingScreen}px)`,
+    }),
+    isPhone: useMediaQuery({query: `(max-width: ${belowTabletScreen}px)`}),
+    isPortrait: useMediaQuery({query: `(orientation: ${ORIENTATIONS.portrait})`}),
+    isLandscape: useMediaQuery({query: `(orientation: ${ORIENTATIONS.landscape})`}),
+  };
+}